% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Retrieve_answer.R
\name{retrieve_answer}
\alias{retrieve_answer}
\title{R DATA ANALYSIS}
\usage{
retrieve_answer(exercise_num)
}
\arguments{
\item{exercise_num}{El número del ejercicio (entre 1 y 6).}
}
\value{
La solución para el ejercicio especificado.
}
\description{
R DATA ANALYSIS
}
\examples{
retrieve_answer(2)
retrieve_answer(5)
}
